requirements.txt
# Core
streamlit>=1.28.0

# LLM Providers
openai>=1.0.0
anthropic>=0.18.0
groq>=0.4.0
google-genai>=1.0.0
requests>=2.28.0
